chore: move job-discovery related code to one module
