+++ packages/apps/human-app/frontend/src/modules/worker/jobs-discovery/components/index.ts
@@ -1,0 +1,4 @@
+export * from './oracles-table-job-types-select';
+export * from './oracles-table';
+export * from './oracles-table-item-mobile';
+export * from './oracles-table-desktop';
