+++ packages/apps/human-app/frontend/src/modules/worker/jobs-discovery/helpers/handle-oracle-navigation.ts
@@ -1,10 +1,10 @@
 import { type NavigateFunction } from 'react-router-dom';
-import { shouldNavigateToRegistration } from '@/modules/worker/components/jobs-discovery/helpers/should-navigate-to-registration';
-import { navigateToRegistrationPage } from '@/modules/worker/components/jobs-discovery/helpers/navigate-to-registration-page';
-import { isHCaptchaOracle } from '@/modules/worker/components/jobs-discovery/helpers/is-hcaptcha-oracle';
-import type { Oracle } from '@/modules/worker/services/oracles';
-import { navigateToHCaptchaPage } from '@/modules/worker/components/jobs-discovery/helpers/navigate-to-hcaptcha-page';
-import { navigateToOracleJobsPage } from '@/modules/worker/components/jobs-discovery/helpers/navigate-to-oracle-jobs-page';
+import { type Oracle } from '../hooks';
+import { shouldNavigateToRegistration } from './should-navigate-to-registration';
+import { navigateToRegistrationPage } from './navigate-to-registration-page';
+import { isHCaptchaOracle } from './is-hcaptcha-oracle';
+import { navigateToHCaptchaPage } from './navigate-to-hcaptcha-page';
+import { navigateToOracleJobsPage } from './navigate-to-oracle-jobs-page';
 
 interface RegistrationResult {
   oracle_addresses: string[];
