+++ packages/apps/human-app/frontend/src/modules/worker/jobs-discovery/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './handle-oracle-navigation';
