+++ packages/apps/human-app/frontend/src/modules/worker/jobs-discovery/helpers/navigate-to-registration-page.ts
