+++ packages/apps/human-app/frontend/src/modules/worker/jobs-discovery/hooks/use-get-oracles.ts
@@ -46,7 +46,7 @@ const H_CAPTCHA_ORACLE: Oracle = {
   registrationNeeded: false,
 };
 
-export async function getOracles({
+async function getOracles({
   selected_job_types,
   signal,
 }: {
