+++ packages/apps/human-app/frontend/src/modules/worker/jobs-discovery/index.ts
@@ -1,0 +1,2 @@
+export * from './jobs-discovery.page';
+export * from './hooks/use-get-oracles';
