+++ packages/apps/human-app/frontend/src/router/routes.tsx
@@ -5,7 +5,6 @@ import { SendResetLinkWorkerSuccessPage } from '@/modules/worker/views/send-rese
 import { ResetPasswordWorkerPage } from '@/modules/worker/views/reset-password/reset-password.page';
 import { SendResetLinkWorkerPage } from '@/modules/worker/views/send-reset-link/send-reset-link.page';
 import { ResetPasswordWorkerSuccessPage } from '@/modules/worker/views/reset-password/reset-password-success.page';
-import { JobsDiscoveryPage } from '@/modules/worker/views/jobs-discovery/jobs-discovery.page';
 import { JobsPage } from '@/modules/worker/views/jobs/jobs.page';
 import { env } from '@/shared/env';
 import { RegistrationPage } from '@/modules/worker/oracle-registration';
@@ -33,6 +32,7 @@ import {
 } from '@/modules/worker/hcaptcha-labeling';
 import { SignUpWorkerPage } from '@/modules/signup/worker';
 import { SignInWorkerPage } from '@/modules/signin/worker';
+import { JobsDiscoveryPage } from '@/modules/worker/jobs-discovery';
 
 export const unprotectedRoutes: RouteProps[] = [
   {
